+++ src/tekore/_client/__init__.py
@@ -1,1 +1,2 @@
 from .full import Spotify
+from .short_link import is_short_link